+++ Frontend/src/components/home/DailyWeather.tsx
