+++ Frontend/src/components/home/WeeklyWeather.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { Text } from "react-native";
+
+const DisplayWeeklyWeather = () => {
+  return <Text>New File</Text>;
+};
+
+export default DisplayWeeklyWeather;
